Add message waiting indicator
One of the nice features about SIP phones is that they can subscribe to
the voicemail box and tell you if anything is in it.  This code adds a
standard messages-summary subscription to sipdroid and keys the
voicemail notifier off it (if the sip system supports, it will also tell
you how many messages you have waiting).  If the returned NOTIFY also gives
a voicemail number to call, this will be called on pressing the notification
otherwise you will get an empty dialer.

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>



git-svn-id: http://sipdroid.googlecode.com/svn/trunk@331 8b64e154-0f13-11de-9106-f38757f67524